+++ lib/mondrian/olap/query.rb
@@ -209,7 +209,9 @@ module Mondrian
       end
 
       def execute
-        @connection.execute to_mdx
+        Error.wrap_native_exception do
+          @connection.execute to_mdx
+        end
       end
 
       private
@@ -255,7 +257,10 @@ module Mondrian
       }
 
       def members_to_mdx(members)
-        if members.length == 1
+        # if only one member which does not end with ]
+        # then assume it is expression which returns set
+        # TODO: maybe always include also single expressions in {...} to avoid some edge cases?
+        if members.length == 1 && members[0][-1,1] != ']'
           members[0]
         elsif members[0].is_a?(Symbol)
           case members[0]
@@ -302,13 +307,13 @@ module Mondrian
         if @where[0].is_a?(Symbol) ||
             @where.length > 1 && @where.map{|full_name| extract_dimension_name(full_name)}.uniq.length == 1
           members_to_mdx(@where)
-        # generate tupple MDX expression
+        # generate tuple MDX expression
         else
-          where_to_mdx_tupple
+          where_to_mdx_tuple
         end
       end
 
-      def where_to_mdx_tupple
+      def where_to_mdx_tuple
         mdx = '('
         mdx << @where.map do |condition|
           condition
@@ -330,7 +335,7 @@ module Mondrian
       end
 
       def extract_dimension_name(full_name)
-        if full_name =~ /\A\[([^\]]+)\]/
+        if full_name =~ /\A[^\[]*\[([^\]]+)\]/
           $1
         end
       end