+++ lib/mondrian/olap/schema.rb
@@ -20,7 +20,7 @@ module Mondrian
         name, attributes = self.class.pre_process_arguments(name, attributes)
         pre_process_attributes(attributes)
         @attributes[:name] = name || @attributes[:name] || 'default' # otherwise connection with empty name fails
-        instance_eval &block if block
+        instance_eval(&block) if block
         self
       end
 
@@ -50,7 +50,7 @@ module Mondrian
       public
 
       attributes :name, :description
-      elements :dimension, :cube, :user_defined_function
+      elements :dimension, :cube, :role, :user_defined_function
 
       class Cube < SchemaElement
         attributes :name, :description,
@@ -76,6 +76,7 @@ module Mondrian
 
       class View < SchemaElement
         attributes :alias
+        data_dictionary_names :alias
         # Defines a "table" using SQL query which can have different variants for different underlying databases
         elements :sql
       end
@@ -127,9 +128,9 @@ module Mondrian
       end
 
       class Join < SchemaElement
-        attributes :left_key, :right_key
-        data_dictionary_names :left_key, :right_key # values in XML will be uppercased when using Oracle driver
-        elements :table
+        attributes :left_key, :right_key, :left_alias, :right_alias
+        data_dictionary_names :left_key, :right_key, :left_alias, :right_alias # values in XML will be uppercased when using Oracle driver
+        elements :table, :join
       end
 
       class Level < SchemaElement
@@ -155,11 +156,11 @@ module Mondrian
           # When generating SQL statements, Mondrian encloses values for String columns in quotation marks,
           # but leaves values for Integer and Numeric columns un-quoted.
           # Date, Time, and Timestamp values are quoted according to the SQL dialect.
-          # For a SQL-compliant dialect, the values appear prefixed by their typename, 
+          # For a SQL-compliant dialect, the values appear prefixed by their typename,
           # for example, "DATE '2006-06-01'".
           # Default value: 'String'
           :type,
-          # Whether members are unique across all parents. 
+          # Whether members are unique across all parents.
           # For example, zipcodes are unique across all states.
           # The first level's members are always unique.
           # Default value: false
@@ -176,7 +177,10 @@ module Mondrian
           # IfBlankName (a member doesn't appear if its name is null, empty or all whitespace);
           # and IfParentsName (a member appears unless its name matches the parent's.
           # Default value: 'Never'
-          :hide_member_if
+          :hide_member_if,
+          # The estimated number of members in this level. Setting this property improves the performance of
+          # MDSCHEMA_LEVELS, MDSCHEMA_HIERARCHIES and MDSCHEMA_DIMENSIONS XMLA requests
+          :approx_row_count
         data_dictionary_names :table, :column, :name_column, :ordinal_column, :parent_column # values in XML will be uppercased when using Oracle driver
         elements :key_expression, :name_expression, :ordinal_expression, :member_formatter, :property
       end
@@ -211,6 +215,7 @@ module Mondrian
           # (if the database permits columns in the SELECT that are not in the GROUP BY).
           # This can be a significant performance enhancement on some databases, such as MySQL.
           :depends_on_level_value
+        data_dictionary_names :column
         elements :property_formatter
       end
 
@@ -300,6 +305,81 @@ module Mondrian
       class AggExclude < SchemaElement
         attributes :name, :pattern, :ignorecase
         data_dictionary_names :name, :pattern
+      end
+
+      class Role < SchemaElement
+        attributes :name
+        elements :schema_grant, :union
+      end
+
+      class SchemaGrant < SchemaElement
+        # access may be "all", "all_dimensions", "custom" or "none".
+        # If access is "all_dimensions", the role has access to all dimensions but still needs explicit access to cubes.
+        # If access is "custom", no access will be inherited by cubes for which no explicit rule is set.
+        # If access is "all_dimensions", an implicut access is given to all dimensions of the schema's cubes,
+        # provided the cube's access attribute is either "custom" or "all"
+        attributes :access
+        elements :cube_grant
+      end
+
+      class CubeGrant < SchemaElement
+        # access may be "all", "custom", or "none".
+        # If access is "custom", no access will be inherited by the dimensions of this cube,
+        # unless the parent SchemaGrant is set to "all_dimensions"
+        attributes :access,
+          # The unique name of the cube
+          :cube
+        elements :dimension_grant, :hierarchy_grant
+      end
+
+      class DimensionGrant < SchemaElement
+        # access may be "all", "custom" or "none".
+        # Note that a role is implicitly given access to a dimension when it is given "all" acess to a cube.
+        # If access is "custom", no access will be inherited by the hierarchies of this dimension.
+        # If the parent schema access is "all_dimensions", this timension will inherit access "all".
+        # See also the "all_dimensions" option of the "SchemaGrant" element.
+        attributes :access,
+          # The unique name of the dimension
+          :dimension
+      end
+
+      class HierarchyGrant < SchemaElement
+        # access may be "all", "custom" or "none".
+        # If access is "custom", you may also specify the attributes :top_level, :bottom_level, and the member grants.
+        # If access is "custom", the child levels of this hierarchy will not inherit access rights from this hierarchy,
+        # should there be no explicit rules defined for the said child level.
+        attributes :access,
+          # The unique name of the hierarchy
+          :hierarchy,
+          # Unique name of the highest level of the hierarchy from which this role is allowed to see members.
+          # May only be specified if the HierarchyGrant.access is "custom".
+          # If not specified, role can see members up to the top level.
+          :top_level,
+          # Unique name of the lowest level of the hierarchy from which this role is allowed to see members.
+          # May only be specified if the HierarchyGrant.access is "custom".
+          # If not specified, role can see members down to the leaf level.
+          :bottom_level,
+          # Policy which determines how cell values are calculated if not all of the children of the current cell
+          # are visible to the current role.
+          # Allowable values are "full" (the default), "partial", and "hidden".
+          :rollup_policy
+        elements :member_grant
+      end
+
+      class MemberGrant < SchemaElement
+        # The children of this member inherit that access.
+        # You can implicitly see a member if you can see any of its children.
+        attributes :access,
+          # The unique name of the member
+          :member
+      end
+
+      class Union < SchemaElement
+        elements :role_usage
+      end
+
+      class RoleUsage < SchemaElement
+        attributes :role_name
       end
 
     end